backend/requirements.txt
# =============================================================================
# MegaStore Backend Dependencies
# =============================================================================

# Django Core
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2
django-extensions==3.2.3

# Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Caching & Task Queue
redis==5.0.4
django-redis==5.4.0
celery==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1

# Search
elasticsearch==8.12.1
elasticsearch-dsl==8.12.0
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5

# Payments
stripe==8.8.0

# File Handling
Pillow==10.3.0
django-storages==1.14.2
boto3==1.34.69

# Serialization & Validation
drf-writable-nested==0.7.0
phonenumbers==8.13.33
django-phonenumber-field==7.3.0

# API Documentation
drf-spectacular==0.27.2

# Security
django-ratelimit==4.1.0
django-csp==3.8

# Server
gunicorn==22.0.0
uvicorn==0.29.0
whitenoise==6.6.0

# Utilities
python-decouple==3.8
python-dotenv==1.0.1

# Code Quality (dev)
flake8==7.0.0
black==24.3.0
isort==5.13.2
mypy==1.9.0
django-stubs==4.2.7

# Testing (dev)
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
faker==24.4.0
model-bakery==1.17.0

# Monitoring
sentry-sdk==1.45.0
django-health-check==3.18.1

# Logging
structlog==24.1.0
